♻️ refactor: 도메인별 테스트 코드 이름 변경
